vsr: head can be recovered in an idle cluster
When we are in `.recovering_head`, the `op == message.header.op` case
isn't necessary outdated, it tells us that our provisional op is in fact
ok. This is especially important in idle cluster, where we might not get
a fresher op.

We _could_ return here if are not in .recovering_head, but let's rather
hit the rest of assertions.