vopr: catch replica-local bugs earlier
State checker catches bug once they break cluster-wide linearizability.
However not every correctness bug immediately leads to cluster divergance.

For example, if a replica forgets its prepare_ok, this might lead to
erroneous truncation during a view change, but it is more likely to be
just repaired over.

In this PR, we enhance the state checker to take a closer look at
individual state of each replica, and prevent replicas from rolling back
their commitments.

Note that we can't track this purely with asserts inside the replica, as
the invariant might be broken across restarts.

I've verified that it makes discovering problem in

https://github.com/tigerbeetledb/tigerbeetle/pull/828

more likely.